reorganize unit tests in running groups
