+++ sdk/core/core/test/cmocka/test_az_policy.c
@@ -34,10 +34,10 @@ az_result test_policy_transport_retry_response_with_header_2(
     void* p_options,
     _az_http_request* p_request,
     az_http_response* p_response);
-void test_az_http_pipeline_policy_credential();
-void test_az_http_pipeline_policy_retry();
-void test_az_http_pipeline_policy_retry_with_header();
-void test_az_http_pipeline_policy_retry_with_header_2();
+void test_az_http_pipeline_policy_credential(void** state);
+void test_az_http_pipeline_policy_retry(void** state);
+void test_az_http_pipeline_policy_retry_with_header(void** state);
+void test_az_http_pipeline_policy_retry_with_header_2(void** state);
 #endif // MOCK_ENABLED
 
 static az_result test_policy_transport(
@@ -46,26 +46,9 @@ static az_result test_policy_transport(
     _az_http_request* p_request,
     az_http_response* p_response);
 
-void test_az_http_pipeline_policy_apiversion();
-void test_az_http_pipeline_policy_uniquerequestid();
-void test_az_http_pipeline_policy_telemetry();
-
-void test_az_http_policy(void** state)
-{
-  (void)state;
-
-/* Tests using wrap to mock. Only suported by gcc */
-#ifdef MOCK_ENABLED
-  test_az_http_pipeline_policy_credential();
-  test_az_http_pipeline_policy_retry();
-  test_az_http_pipeline_policy_retry_with_header();
-  test_az_http_pipeline_policy_retry_with_header_2();
-#endif // MOCK_ENABLED
-
-  test_az_http_pipeline_policy_apiversion();
-  test_az_http_pipeline_policy_uniquerequestid();
-  test_az_http_pipeline_policy_telemetry();
-}
+void test_az_http_pipeline_policy_apiversion(void** state);
+void test_az_http_pipeline_policy_uniquerequestid(void** state);
+void test_az_http_pipeline_policy_telemetry(void** state);
 
 az_result test_policy_transport(
     _az_http_policy* p_policies,
@@ -80,8 +63,10 @@ az_result test_policy_transport(
   return AZ_OK;
 }
 
-void test_az_http_pipeline_policy_telemetry()
+void test_az_http_pipeline_policy_telemetry(void** state)
 {
+  (void)state;
+
   uint8_t buf[100];
   uint8_t header_buf[(2 * sizeof(az_pair))];
   memset(buf, 0, sizeof(buf));
@@ -112,8 +97,10 @@ void test_az_http_pipeline_policy_telemetry()
   assert_return_code(az_http_pipeline_policy_telemetry(policies, &telemetry, &hrb, NULL), AZ_OK);
 }
 
-void test_az_http_pipeline_policy_apiversion()
+void test_az_http_pipeline_policy_apiversion(void** state)
 {
+  (void)state;
+
   uint8_t buf[100];
   uint8_t header_buf[(2 * sizeof(az_pair))];
   memset(buf, 0, sizeof(buf));
@@ -151,8 +138,10 @@ void test_az_http_pipeline_policy_apiversion()
   assert_return_code(az_http_pipeline_policy_apiversion(policies, &api_version, &hrb, NULL), AZ_OK);
 }
 
-void test_az_http_pipeline_policy_uniquerequestid()
+void test_az_http_pipeline_policy_uniquerequestid(void** state)
 {
+  (void)state;
+
   uint8_t buf[100];
   uint8_t header_buf[(2 * sizeof(az_pair))];
   memset(buf, 0, sizeof(buf));
@@ -247,8 +236,10 @@ az_result test_policy_transport_retry_response_with_header_2(
   return AZ_OK;
 }
 
-void test_az_http_pipeline_policy_credential()
+void test_az_http_pipeline_policy_credential(void** state)
 {
+  (void)state;
+
   uint8_t buf[100];
   uint8_t header_buf[(2 * sizeof(az_pair))];
   memset(buf, 0, sizeof(buf));
@@ -288,8 +279,10 @@ void test_az_http_pipeline_policy_credential()
   assert_return_code(az_http_pipeline_policy_credential(policies, &credential, &hrb, NULL), AZ_OK);
 }
 
-void test_az_http_pipeline_policy_retry()
+void test_az_http_pipeline_policy_retry(void** state)
 {
+  (void)state;
+
   uint8_t buf[100];
   uint8_t header_buf[(2 * sizeof(az_pair))];
   memset(buf, 0, sizeof(buf));
@@ -327,8 +320,10 @@ void test_az_http_pipeline_policy_retry()
       az_http_pipeline_policy_retry(policies, &retry_options, &hrb, &response), AZ_OK);
 }
 
-void test_az_http_pipeline_policy_retry_with_header()
+void test_az_http_pipeline_policy_retry_with_header(void** state)
 {
+  (void)state;
+
   uint8_t buf[100];
   uint8_t header_buf[(2 * sizeof(az_pair))];
   memset(buf, 0, sizeof(buf));
@@ -366,8 +361,10 @@ void test_az_http_pipeline_policy_retry_with_header()
       az_http_pipeline_policy_retry(policies, &retry_options, &hrb, &response), AZ_OK);
 }
 
-void test_az_http_pipeline_policy_retry_with_header_2()
+void test_az_http_pipeline_policy_retry_with_header_2(void** state)
 {
+  (void)state;
+
   uint8_t buf[100];
   uint8_t header_buf[(2 * sizeof(az_pair))];
   memset(buf, 0, sizeof(buf));
@@ -406,3 +403,19 @@ void test_az_http_pipeline_policy_retry_with_header_2()
 }
 
 #endif // MOCK_ENABLED
+
+int test_az_policy()
+{
+  const struct CMUnitTest tests[] = {
+#ifdef MOCK_ENABLED
+    cmocka_unit_test(test_az_http_pipeline_policy_credential),
+    cmocka_unit_test(test_az_http_pipeline_policy_retry),
+    cmocka_unit_test(test_az_http_pipeline_policy_retry_with_header),
+    cmocka_unit_test(test_az_http_pipeline_policy_retry_with_header_2),
+#endif // MOCK_ENABLED
+    cmocka_unit_test(test_az_http_pipeline_policy_apiversion),
+    cmocka_unit_test(test_az_http_pipeline_policy_uniquerequestid),
+    cmocka_unit_test(test_az_http_pipeline_policy_telemetry),
+  };
+  return cmocka_run_group_tests_name("az_core_policy", tests, NULL, NULL);
+}
